Refactored maps.js and split it into several files
I also tried to rewrite the placement algorithm
the way discussed with Paul, but failed to do so.
I think the problem is, that collisions between bboxes of
different WordOverlays are found even when they should not exist.
I believe this to be due to wrong bboxes,
but couldn't figure out where it comes from.